Fix byte underflow infinite loop in AppearanceManager.UpdateBones
When bones.Count is 0, the byte loop variable i underflows from 0 to
255 after removal, causing an infinite loop since 255 >= 0 is always
true. Changed loop variable from byte to int so i becomes -1 on
underflow and the loop exits correctly.

+++ Source/NexusForever.Game/Entity/AppearanceManager.cs
@@ -166,9 +166,9 @@ namespace NexusForever.Game.Entity
                     .ToList()
             }, true);
 
-            for (byte i = (byte)characterBones.Count; i >= bones.Count; i--)
+            for (int i = characterBones.Count; i >= bones.Count; i--)
             {
-                if (!characterBones.Remove(i, out IBone bone))
+                if (!characterBones.Remove((byte)i, out IBone bone))
                     continue;
 
                 bone.Delete();
